requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiobotocore==2.24.2
    # via s3fs
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via
    #   aiobotocore
    #   litellm
    #   s3fs

firebase-admin
google-auth
prefect
aioitertools==0.12.0
    # via aiobotocore
aiosignal==1.4.0
    # via aiohttp
alembic==1.16.5
    # via optuna
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   asyncer
    #   dspy
    #   httpx
    #   openai
asyncer==0.0.8
    # via dspy
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via dspy
base58==2.1.1
    # via
    #   scalecodec
    #   substrate-interface
botocore==1.40.18
    # via aiobotocore
cachetools==6.2.0
    # via dspy
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
    #   substrate-interface
cffi==1.17.1
    # via pynacl
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via litellm
cloudpickle==3.1.1
    # via dspy
colorlog==6.9.0
    # via optuna
cytoolz==1.0.1
    # via eth-utils
diskcache==5.6.3
    # via dspy
distro==1.9.0
    # via openai
dspy==3.0.3
    # via -r requirements.in
ecdsa==0.19.1
    # via substrate-interface
eth-hash==0.7.1
    # via eth-utils
eth-keys==0.7.0
    # via substrate-interface
eth-typing==5.2.1
    # via
    #   eth-keys
    #   eth-utils
eth-utils==5.3.1
    # via
    #   eth-keys
    #   substrate-interface
fastuuid==0.12.0
    # via litellm
filelock==3.19.1
    # via huggingface-hub
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.7.0
    # via
    #   huggingface-hub
    #   s3fs
gepa[dspy]==0.0.7
    # via dspy
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via httpcore
hf-xet==1.1.9
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   litellm
    #   openai
huggingface-hub==0.34.4
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   substrate-interface
    #   yarl
importlib-metadata==8.7.0
    # via litellm
jinja2==3.1.6
    # via litellm
jiter==0.10.0
    # via openai
jmespath==1.0.1
    # via
    #   aiobotocore
    #   botocore
joblib==1.5.2
    # via dspy
json-repair==0.50.1
    # via dspy
jsonschema==4.25.1
    # via litellm
jsonschema-specifications==2025.9.1
    # via jsonschema
litellm==1.76.3
    # via dspy
magicattr==0.1.6
    # via dspy
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.8.0
    # via scalecodec
multidict==6.6.4
    # via
    #   aiobotocore
    #   aiohttp
    #   yarl
numpy==2.3.2
    # via
    #   dspy
    #   optuna
openai==1.106.1
    # via
    #   dspy
    #   litellm
optuna==4.5.0
    # via dspy
orjson==3.11.3
    # via dspy
packaging==25.0
    # via
    #   huggingface-hub
    #   optuna
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
py-bip39-bindings==0.3.0
    # via substrate-interface
py-ed25519-zebra-bindings==1.3.0
    # via substrate-interface
py-sr25519-bindings==0.2.2
    # via substrate-interface
pycparser==2.22
    # via cffi
pycryptodome==3.23.0
    # via substrate-interface
pydantic==2.11.7
    # via
    #   dspy
    #   eth-utils
    #   litellm
    #   openai
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pynacl==1.5.0
    # via substrate-interface
python-dateutil==2.9.0.post0
    # via
    #   aiobotocore
    #   botocore
python-dotenv==1.1.1
    # via litellm
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   optuna
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.9.1
    # via
    #   dspy
    #   tiktoken
requests==2.32.5
    # via
    #   dspy
    #   huggingface-hub
    #   scalecodec
    #   substrate-interface
    #   tiktoken
rich==14.1.0
    # via dspy
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
s3fs==2025.7.0
    # via -r requirements.in
scalecodec==1.2.11
    # via substrate-interface
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.43
    # via
    #   alembic
    #   optuna
substrate-interface==1.7.11
    # via -r requirements.in
tenacity==9.1.2
    # via dspy
tiktoken==0.11.0
    # via litellm
tokenizers==0.22.0
    # via litellm
toolz==1.0.0
    # via cytoolz
tqdm==4.67.1
    # via
    #   dspy
    #   huggingface-hub
    #   openai
    #   optuna
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   alembic
    #   anyio
    #   eth-typing
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via
    #   botocore
    #   requests
websocket-client==1.8.0
    # via substrate-interface
wrapt==1.17.3
    # via aiobotocore
xxhash==3.5.0
    # via
    #   dspy
    #   substrate-interface
yarl==1.20.1
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
